% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ECoF_releases.R
\name{available_ECoF_releases}
\alias{available_ECoF_releases}
\title{List available releases of the Eschmeyer's Catalog of Fishes}
\usage{
available_ECoF_releases()
}
\value{
A character vector
}
\description{
Lists all available releases (year.month format).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
available_ECoF_releases()
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
