% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_taxonomy.R
\name{build_fish_taxonomy}
\alias{build_fish_taxonomy}
\title{Build fish taxonomy}
\usage{
build_fish_taxonomy(
  names,
  id.rank = NULL,
  check_names = TRUE,
  colname = "original_name",
  db = c("FB", "ECoF"),
  version = "latest"
)
}
\arguments{
\item{names}{A character vector of taxonomic names. The function accepts
both species names and names of fish identified at higher ranks}

\item{id.rank}{A character vector of taxonomic ranks describing the level
at which taxa are identified. Must be the same length as 'names'.
Accepted ranks are: "species", "genus", "family", "order", and "class".
If NULL (default), the function tries to guess the ranks from the names}

\item{check_names}{Logical, check whether taxonomic names are correct at
the respective taxonomic rank. Default to TRUE}

\item{colname}{A name for the column containing the supplied names
in the returned data frame}

\item{db}{Either "FB" (FishBase; default) or "ECoF" (Eschmeyer's Catalog of Fishes)}

\item{version}{A string; the database version to use. Will default to the latest
available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}} for FishBase
and \code{\link[=available_ECoF_releases]{available_ECoF_releases()}} for the Eschmeyer's Catalog of Fishes}
}
\value{
A data frame including the supplied names, ranks and taxonomy
}
\description{
Build taxonomy for a list of fish taxa identified at different taxonomic ranks
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
species <- c("Dentex dentex", "Mullus sp.", "Balistidae spp.",
             "Carangiformes spp.", "Elasmobranchii spp.")
# Declaring level of identification
ranks = c("species", "genus", "family", "order", "class")
build_fish_taxonomy(species, id.rank = ranks)
# Letting the function guess the level of identification
build_fish_taxonomy(species)
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
