% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trait_table.R
\name{build_trait_table}
\alias{build_trait_table}
\title{Build a fish trait table}
\usage{
build_trait_table(
  names,
  fns,
  args = vector("list", length(fns)),
  FB_version = latest_release("FB")
)
}
\arguments{
\item{names}{A data frame created by \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}} or a
character vector of taxonomic names to be passed on to \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}}}

\item{fns}{List of \code{fishrfun} functions that return species-level traits}

\item{args}{List of same length as \code{fns}. Each element must be a list of
arguments, other than "names" and "FB_version", to be passed on to the
respective function in \code{fns}. The element must be an empty list if no
additional arguments are required. See \emph{Examples}}

\item{FB_version}{A string; the FishBase version to use.
Will default to the latest available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}}}
}
\value{
A data frame
}
\description{
A wrapper function to build a table with multiple traits extracted from FishBase
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# List of functions for the traits of interest
fns <- list(aspect_ratio_FB, max_length_FB, trophic_level_FB)
# Pass an additional argument to `max_length_FB` and `trophic_level_FB`
args <- list(list(), list(length_type = "both"), list(type = "food items"))
# List of taxa identified to species, genus or family level
species <- c("Boops boops", "Serranus sp.", "Sparidae spp.")
ranks <- c("species", "genus", "family")
taxa <- build_fish_taxonomy(species, ranks)
build_trait_table(taxa, fns, args)
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
