% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_phylogeny.R
\name{build_tree_from_taxonomy}
\alias{build_tree_from_taxonomy}
\title{Convert taxonomy into a phylogenetic tree}
\usage{
build_tree_from_taxonomy(taxonomy, formula = ~class/order/family/genus/species)
}
\arguments{
\item{taxonomy}{A data frame with one column for each taxonomic variable in \code{formula}}

\item{formula}{A right-side formula describing the taxonomic relationship.
See \code{\link[ape:as.phylo.formula]{ape::as.phylo.formula()}} for details}
}
\value{
An object of class phylo
}
\description{
\code{build_tree_from_taxonomy} is a wrapper around \code{\link[ape:as.phylo.formula]{ape::as.phylo.formula()}}
and converts taxonomic hierarchy into a phylogenetic tree.
Following \href{https://doi.org/10.1111/2041-210X.14076}{Thorson et al. 2023},
all edge lengths are set to 1.
}
\examples{
taxa <- tibble::tribble(
  ~"species", ~"genus", ~"family", ~"order", ~"class",
  "sp1",      "gen1",   "fam1",    "ord1",   "cla1",
  "sp2",      "gen2",   "fam1",    "ord1",   "cla1",
  "sp3",      "gen3",   "fam2",    "ord1",   "cla1",
  "fam3 sp.", "",       "fam3",    "ord2",   "cla1",
  "sp5",      "gen5",   "fam4",    "ord2",   "cla1",
  "sp6",      "gen6",   "fam5",    "ord3",   "cla1",
  "sp7",      "gen7",   "fam6",    "ord4",   "cla1",
  "gen7 sp.", "gen7",   "fam6",    "ord4",   "cla1"
)
tree <- build_tree_from_taxonomy(taxa)
tree
plot(tree)
}
\references{
Thorson J. T. et al. (2023) Identifying direct and indirect associations
among traits by merging phylogenetic comparative methods and structural
equation models. Methods Ecol Evol, 14(5): 1259-1275.
\url{https://doi.org/10.1111/2041-210X.14076}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
