% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fish_names.R
\name{check_fish_names}
\alias{check_fish_names}
\title{Check fish scientific names}
\usage{
check_fish_names(
  names,
  rank = "Species",
  db = c("FB", "ECoF"),
  version = "latest"
)
}
\arguments{
\item{names}{A string or character vector of names}

\item{rank}{A string; the taxonomic rank at which names should be checked.
Must be one of "Species" (default), "Genus", "Family", "Order", or "Class"}

\item{db}{Either "FB" (FishBase; default) or "ECoF" (Eschmeyer's Catalog of Fishes)}

\item{version}{A string; the database version to use. Will default to the latest
available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}} for FishBase
and \code{\link[=available_ECoF_releases]{available_ECoF_releases()}} for the Eschmeyer's Catalog of Fishes}
}
\value{
A message. If any non-valid name is found, a data frame is also returned
invisibly, including non-valid supplied names and suggested names.
If \code{rank = "Species"}, the data frame also includes valid names returned by
\code{\link[rfishbase:validate_names]{rfishbase::validate_names()}} or \code{\link[=validate_names_ECoF]{validate_names_ECoF()}}.
}
\description{
Check fish scientific names
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Correct names in FishBase
check_fish_names(c("Boops boops", "Sparus aurata"))
# Misspelled family name and suggested name
# using the Eschmeyer's Catalog of Fishes
errors <- check_fish_names(c("Balistidae", "Seranidae"),
                           rank = "Family",
                           db = "ECoF")
errors
# Non-valid species name, suggested name and valid name
errors <- check_fish_names(c("Scarus ghobban", "Ostracion cubicus"))
errors
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
