% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_params_FB.R
\name{lw_params_FB}
\alias{lw_params_FB}
\title{Get length-weight parameters from FishBase}
\usage{
lw_params_FB(names, FB_version = latest_release("FB"))
}
\arguments{
\item{names}{A data frame created by \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}} or a
character vector of taxonomic names to be passed on to \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}}}

\item{FB_version}{A string; the FishBase version to use.
Will default to the latest available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}}}
}
\value{
A data frame with the provided name(s), means and standard deviations
of \emph{a} and \emph{b}, and the method used. Further information about the applied
method can be found in \href{https://doi.org/10.1111/jai.12299}{Froese et al. 2014}
and \href{https://fishbase.se/fishonline/english/fol_fishbasegoesfishbayes.htm}{here}
on FishBase.
}
\description{
\code{lw_params_FB} extracts the \emph{a} and \emph{b} parameters of the length-weight
relationship (\eqn{W = a * L^b}) as estimated by
\href{https://doi.org/10.1111/jai.12299}{Froese et al. 2014}.
It works with taxa identified at different taxonomic ranks.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Length-weight parameters of fishes identified to species, genus or family level
species <- c("Dentex dentex", "Labrus sp.", "Scombridae spp.")
ranks <- c("species", "genus", "family")
taxa <- build_fish_taxonomy(species, ranks)
lw_params_FB(taxa)
\dontshow{\}) # examplesIf}
}
\references{
Froese R., Thorson J. T. and Reyes R. B. Jr. (2014) A Bayesian approach for
estimating length-weight relationships in fishes. J Appl Ichthyol, 30(1):78-85
\url{https://doi.org/10.1111/jai.12299}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
