% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_length_FB.R
\name{max_length_FB}
\alias{max_length_FB}
\title{Get maximum length from FishBase}
\usage{
max_length_FB(
  names,
  length_type = c("TL", "SL", "both"),
  FB_version = latest_release("FB")
)
}
\arguments{
\item{names}{A data frame created by \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}} or a
character vector of taxonomic names to be passed on to \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}}}

\item{length_type}{A string; the type of length to return.
"TL" (default) or "SL", for total and standard length respectively,
or "both" if both are to be returned}

\item{FB_version}{A string; the FishBase version to use.
Will default to the latest available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}}}
}
\value{
A data frame with the provided name(s) and maximum length(s).
}
\description{
\code{max_length_FB} extracts the observed or estimated species maximum length
(total, standard, or both). It also works with taxa identified at higher
taxonomic ranks. In this case it returns the maximum length among
all the species belonging to the rank.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# List of taxa identified to species, genus or family level
species <- c("Dentex dentex", "Labrus sp.", "Scombridae spp.")
ranks <- c("species", "genus", "family")
taxa <- build_fish_taxonomy(species, ranks)
# Max total length of fishes identified to species, genus or family level
max_length_FB(taxa)
# Both max TL and SL
max_length_FB(taxa, length_type = "both")
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
