% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients_FB.R
\name{nutrients_FB}
\alias{nutrients_FB}
\title{Get nutrients estimates from FishBase}
\usage{
nutrients_FB(names, FB_version = latest_release("FB"))
}
\arguments{
\item{names}{A data frame created by \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}} or a
character vector of taxonomic names to be passed on to \code{\link[=build_fish_taxonomy]{build_fish_taxonomy()}}}

\item{FB_version}{A string; the FishBase version to use.
Will default to the latest available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}}}
}
\value{
A data frame with the provided name(s), medians and 95\% Highest
Posterior Densities of seven nutrients, the taxonomic rank at which the values are extracted
and a column ("method") describing how the values were obtained.
}
\description{
\code{nutrients_FB} extracts the concentrations of seven key nutrients in fish:
calcium (mg per 100 g), iron (mg per 100 g), omega-3 fatty acids (g per 100 g),
protein (\%), selenium (μg per 100 g), vitamin A (mg per 100 g), and
zinc (mg per 100 g). These are estimated by a Bayesian model (available
\href{https://github.com/mamacneil/NutrientFishbase}{here}) which includes
phylogenetic and traits information, an update of the model in
\href{https://doi.org/10.1038/s41586-019-1592-6}{Hicks et al. 2019}.
It works with taxa identified at different taxonomic ranks.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Nutrients concentrations of fishes identified to species, genus or family level
species <- c("Scarus ghobban", "Caranx sp.", "Lutjanidae spp.")
ranks <- c("species", "genus", "family")
taxa <- build_fish_taxonomy(species, ranks)
nutrients_FB(taxa)
\dontshow{\}) # examplesIf}
}
\references{
Hicks C et al. (2019) Harnessing global fisheries to tackle micronutrient
deficiencies. Nature, 574: 95-98. \url{https://doi.org/10.1038/s41586-019-1592-6}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
