% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fish_names.R
\name{validate_names_ECoF}
\alias{validate_names_ECoF}
\title{Validate species names in the Eschmeyer's Catalog of Fishes}
\usage{
validate_names_ECoF(
  species_list,
  USE.NAMES = FALSE,
  version = latest_release("ECoF")
)
}
\arguments{
\item{species_list}{A string or character vector of species names}

\item{USE.NAMES}{Logical; if TRUE use \code{species_list} as names for the
returned object}

\item{version}{A string; the database version to use. Will default to
the latest available; see \code{\link[=available_ECoF_releases]{available_ECoF_releases()}}}
}
\value{
A vector of validated names, or a list if multiple
valid species match a single name
}
\description{
Validate species names in the Eschmeyer's Catalog of Fishes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
validate_names_ECoF("Ostracion cubicus")
\dontshow{\}) # examplesIf}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
