% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_taxonomy.R
\name{load_fish_taxonomy}
\alias{load_fish_taxonomy}
\title{Load fish taxonomy}
\usage{
load_fish_taxonomy(db = c("FB", "ECoF"), version = "latest")
}
\arguments{
\item{db}{Either "FB" (FishBase; default) or "ECoF" (Eschmeyer's Catalog of Fishes)}

\item{version}{A string; the database version to use. Will default to the latest
available; see \code{\link[rfishbase:available_releases]{rfishbase::available_releases()}} for FishBase
and \code{\link[=available_ECoF_releases]{available_ECoF_releases()}} for the Eschmeyer's Catalog of Fishes}
}
\value{
A data frame
}
\description{
Load fish taxonomy
}
\examples{
\dontrun{
load_fish_taxonomy()
load_fish_taxonomy("ECoF")
}
}
\author{
Mattia Ghilardi, \email{mattia.ghilardi91@gmail.com}
}
